Improves Writing and Removes References to "Functionalities".
I am not sure if there is a concept of "functionalities" that is different from applications but, since I have not happened across that information and since this is the "community-apps" repo, I have taken the liberty of removing those references. Feel free to add them back in if they do refer to something other than applications, but in that case it might be good to (1) add some documentation to clarify what "functionalities" are and (2) rename this repo to something that does not specify "apps".