Add "Replica" application.
Specifically for users who do 'Full Body Tracking', this application generates a double of your avatar that only you can see. This is useful to master your avatar moves. Good to have for dancing.

Basically, it is a modified Dopple-ganger app, but able to do the room scale translations inst5ead to be stationary. This is really for 'Full Body Tracking'.